Added a file containing models for each FPGA memory instances used in or1200. The file is in bench/verilog/sim_lib/fpga_memory_primitives.v.
With it, people who change the or1200_defines.v inside of the project structure will still be able to simulate, using house-made models, not from manufacturers. 

minsoc_bench.v had to be extended by the task, init_fpga_memory, to initialize the dual or two port memories instances of or1200. This has to be done based on the type of memory used, so many different instantiations based on definitions. Somehow or1200 expects all memory values to be 0 upon start, so this is necessary. 



git-svn-id: http://opencores.org/ocsvn/minsoc/minsoc/trunk@10 1e56cc09-5d24-4cb8-85b2-5a87126a96e3