minsoc_defines.v had a semicolon at the end of the two reset polarity definitions.
minsoc_top had a signal array for two different signals which was for backward compatibility. The compatible debugging module is so old that there is no reason for keeping it.

Documentation has been updated to better explain how to use the definitions files, minsoc_defines.v and or1200_defines.v. An example for Altera devices has been added too.



git-svn-id: http://opencores.org/ocsvn/minsoc/minsoc/trunk@20 1e56cc09-5d24-4cb8-85b2-5a87126a96e3