Documentation revision 1.1, thanks to Wojciech A. Koszek for many comments on it.
Also updating howto, splitint it in INSTALL, HOWTO, FAQ and synthesis_examples, so it should be more clear now what to do when and not to try too much when you don't need. Like everyone was trying to debug the simulation but didn't even test the regular simulation before. Again thanks to Wojciech A. Koszek for his view on this matter. 



git-svn-id: http://opencores.org/ocsvn/minsoc/minsoc/trunk@32 1e56cc09-5d24-4cb8-85b2-5a87126a96e3